Solution to Calculate Average Temperature of a Week

+++ Calculate Average Temperature of a Week@@67877d61dc107486ad5c274d/code.c
@@ -14,7 +14,7 @@ int main(){
     for(int i = 0; i<a; i++){
         scanf("%s %f", &s[i].name, &s[i].temp);
     }
-    int summ = 0;
+    float summ = 0.0;
     for (int i =0; i<a; i++){
         summ += s[i].temp;
     }
